Do a main refactor - Moved all the class in a way that all the packages are now consistent - Insered deprecation warnings in old classes
