+++ src/main/java/it/polpetta/libris/image/google/GoogleAbstractFactory.java
@@ -2,6 +2,7 @@ package it.polpetta.libris.image.google;
 
 import it.polpetta.libris.contract.IAbstractFactoryReverseSearchProvider;
 import it.polpetta.libris.contract.IQueryBuilder;
+import it.polpetta.libris.image.google.hmtlImageSearch.GoogleImageSearchQueryBuilder;
 
 /**
  * Created by davide on 29/04/17.
