+++ src/main/java/it/polpetta/libris/image/google/contract/IGoogleImageSearcher.java
@@ -1,4 +1,4 @@
-package it.polpetta.libris.image.google;
+package it.polpetta.libris.image.google.contract;
 
 import it.polpetta.libris.contract.ISearcher;
 import it.polpetta.libris.image.google.contract.IGoogleImageSearchResult;
@@ -8,6 +8,8 @@ import java.io.IOException;
 /**
  * Created by davide on 06/05/17.
  */
+
+// FIXME THIS SHOULD BE A ABSTRACT CLASS THAT EXTENDS ABSTRACTURLSEARCHER AND RETURNS AN IGOOGLEIMAGESEARCHRESULT
 public interface IGoogleImageSearcher extends ISearcher {
 
     IGoogleImageSearchResult search() throws IOException;
