+++ src/main/java/it/polpetta/libris/image/google/hmtlImageSearch/searchers/URLFactoryMethodSearcher.java
@@ -1,7 +1,7 @@
-package it.polpetta.libris.image.google.searchers;
+package it.polpetta.libris.image.google.hmtlImageSearch.searchers;
 
 import it.polpetta.libris.utils.Coordinates;
-import it.polpetta.libris.image.google.IGoogleImageSearcher;
+import it.polpetta.libris.image.google.contract.IGoogleImageSearcher;
 import it.polpetta.libris.contract.AbstractFactoryMethodSearcher;
 
 import java.net.URL;
@@ -9,6 +9,8 @@ import java.net.URL;
 /**
  * Created by davide on 28/04/17.
  */
+// FIXME REMOVE ME, I'LL BE SUBSTITUTED SOON
+@Deprecated
 public class URLFactoryMethodSearcher extends AbstractFactoryMethodSearcher {
 
     private URL link = null;
